Create Utility Function Tests
